separate testing of absolute into separate classes to avoid icy error
